Remove unnecessary QualityChecker object - Required some refactoring of tests and code :)
